requirements.txt
# FoxBrain RAG Demo — Dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
openai>=1.0.0
numpy>=1.24.0
PyMuPDF>=1.23.0
pymilvus>=2.4.0
python-multipart>=0.0.6
